+++ app/src/main/java/com/example/eshfeenygraduationproject/eshfeeny/alarm/adapter/AlarmAdapter.kt
@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.example.domain.entity.alarm.Alarm
 import com.example.eshfeenygraduationproject.databinding.AlarmItemViewBinding
+import com.example.eshfeenygraduationproject.eshfeeny.alarm.fragment.AlarmFragmentDirections
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
@@ -35,6 +36,10 @@ class AlarmAdapter : ListAdapter<Alarm, AlarmAdapter.ViewHolder>(AlarmDiffCallba
             itemBinding.timeTextView.text = formattedTime
             itemBinding.MedicineNameTextView.text = alarm.name
             itemBinding.MedicineDescTextView.text = alarm.notes
+
+            itemBinding.cardVew.setOnClickListener {
+
+            }
         }
     }
 }
