+++ app/src/main/java/com/example/eshfeenygraduationproject/eshfeeny/alarm/fragment/EditAlarmFragment.kt
@@ -1,0 +1,29 @@
+package com.example.eshfeenygraduationproject.eshfeeny.alarm.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.eshfeenygraduationproject.R
+import com.example.eshfeenygraduationproject.databinding.FragmentEditAlarmBinding
+
+class EditAlarmFragment : Fragment() {
+
+    private var binding: FragmentEditAlarmBinding? = null
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        binding = FragmentEditAlarmBinding.inflate(inflater)
+
+        return binding?.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        binding = null
+    }
+}
