+++ app/src/main/res/navigation/nav_graph.xml
@@ -45,9 +45,6 @@
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
             app:popExitAnim="@anim/nav_default_pop_exit_anim" />
         <action
-            android:id="@+id/action_homeFragment2_to_searchFragment"
-            app:destination="@id/searchFragment" />
-        <action
             android:id="@+id/action_homeFragment2_to_searchResultsFragment"
             app:destination="@id/searchResultsFragment" />
         <action
@@ -125,6 +122,13 @@
         <action
             android:id="@+id/action_alarmFragment_to_setAlarmFragment"
             app:destination="@id/setAlarmFragment"
+            app:enterAnim="@anim/nav_default_enter_anim"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
+        <action
+            android:id="@+id/action_alarmFragment_to_editAlarmFragment"
+            app:destination="@id/editAlarmFragment"
             app:enterAnim="@anim/nav_default_enter_anim"
             app:exitAnim="@anim/nav_default_exit_anim"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
@@ -231,13 +235,6 @@
         <action
             android:id="@+id/action_detailsFragment_to_brandItemsFragment"
             app:destination="@id/brandItemsFragment"
-            app:enterAnim="@anim/nav_default_enter_anim"
-            app:exitAnim="@anim/nav_default_exit_anim"
-            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
-        <action
-            android:id="@+id/action_detailsFragment_to_searchFragment"
-            app:destination="@id/searchFragment"
             app:enterAnim="@anim/nav_default_enter_anim"
             app:exitAnim="@anim/nav_default_exit_anim"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
@@ -494,26 +491,6 @@
             app:popExitAnim="@anim/nav_default_pop_exit_anim" />
     </fragment>
     <fragment
-        android:id="@+id/searchFragment"
-        android:name="com.example.eshfeenygraduationproject.eshfeeny.search.SearchFragment"
-        android:label="fragment_search"
-        tools:layout="@layout/fragment_search">
-        <action
-            android:id="@+id/action_searchFragment_to_homeFragment2"
-            app:destination="@id/homeFragment2"
-            app:enterAnim="@anim/nav_default_enter_anim"
-            app:exitAnim="@anim/nav_default_exit_anim"
-            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
-        <action
-            android:id="@+id/action_searchFragment_to_detailsFragment"
-            app:destination="@id/detailsFragment"
-            app:enterAnim="@anim/nav_default_enter_anim"
-            app:exitAnim="@anim/nav_default_exit_anim"
-            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
-    </fragment>
-    <fragment
         android:id="@+id/bmiResultFragment"
         android:name="com.example.eshfeenygraduationproject.eshfeeny.bmiandbmr.BmiResultFragment"
         android:label="fragment_bmi_result"
@@ -583,5 +560,21 @@
         <action
             android:id="@+id/action_addPhotoICFragment_to_infoInsuranceCardFragment"
             app:destination="@id/infoInsuranceCardFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/editAlarmFragment"
+        android:name="com.example.eshfeenygraduationproject.eshfeeny.alarm.fragment.EditAlarmFragment"
+        android:label="fragment_edit_alarm"
+        tools:layout="@layout/fragment_edit_alarm" >
+        <action
+            android:id="@+id/action_editAlarmFragment_to_alarmFragment"
+            app:destination="@id/alarmFragment"
+            app:enterAnim="@anim/nav_default_enter_anim"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
+        <argument
+            android:name="alarm"
+            app:argType="com.example.domain.entity.alarm.Alarm" />
     </fragment>
 </navigation>
