+++ domain/src/main/java/com/example/domain/entity/alarm/Alarm.kt
@@ -1,5 +1,9 @@
 package com.example.domain.entity.alarm
 
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
 data class Alarm(
     val _id: String,
     val alarmTime: List<Long>,
@@ -10,4 +14,4 @@ data class Alarm(
     val name: String,
     val notes: String?,
     val repetition: String
-)
+) : Parcelable
